fill aliasrc in link list

+++ src/init/init_alias.c
@@ -6,28 +6,47 @@
 */
 
 #include "markup.h"
+#include "alias.h"
 #include "42sh.h"
 #include <stddef.h>
 #include <stdlib.h>
 #include <string.h>
 #include <stdio.h>
+#include <stdbool.h>
 
 mnode_t *parse_file(char *filepath);
+
+int fill_list_alias(llist_t *alias, mnode_t *node)
+{
+	alias_t *fill_alias = malloc(sizeof(alias_t));
+
+	if (!fill_alias) {
+		return (false);
+	}
+	fill_alias->name = node->label;
+	fill_alias->value = node->data.child->value;
+	list_push_tail(fill_alias, alias);
+	return (true);
+}
 
 llist_t *init_alias(void)
 {
 	llist_t *alias = list_init(NULL);
-	mnode_t *root = parse_file("./42shrc");
+	mnode_t *root = parse_file("42shrc");
 	mnode_t *node = get_node("alias/", root)->data.root->head;
-	if (!node) {
-		printf("zbeub\n");
+	int check = 0;
+
+	if (!root) {
+		return (NULL);
 	}
 	while (node) {
-		printf("-----------------\n");
-		printf("Label [%s]\n", node->label);
-		printf("Value [%s]\n", node->data.child->value);
+		check = fill_list_alias(alias, node);
+		if(!check) {
+			return (NULL);
+		}
 		node = node->next;
 	}
 	destroy_tree(root);
+	free(node);
 	return (alias);
 }
